pyproject.toml
[project]
name = "image-captioning"
version = "0.1.0"
description = "Image Captioning with PyTorch and VGG19"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jupyter>=1.0.0",
    "matplotlib>=3.9.0",
    "nltk>=3.8.1",
    "numpy<2.0.0",
    "pandas>=2.2.2",
    "pillow>=10.3.0",
    "pyarrow>=16.1.0",
    "scikit-learn>=1.5.0",
    "tensorboard>=2.20.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "tqdm>=4.66.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
